phase 3: 64-bit audio works — native bank promotion and three stock-libaudio 64-bit bugs
The 64-bit build now boots with music and SFX synthesizing (PCM profile
matches the 32-bit golden: rms 3368 vs 3368, whine bin clean), and the
autostart Dam run passes with identical state hashes to the audio-off
run. Audio is ON by default on 64-bit now (PORT_64_AUDIO=0 disables).

The ctl bank blob is 32-bit layout promoted in place by alBnkfNew through
native structs — replaced on 64-bit by portBnkfPromote (preprocess_audio):
the five pointer-bearing containers (file/bank/instrument/sound/wavetable)
are rebuilt as native structs; the pointer-free leaves (envelopes,
keymaps, loops, books) stay in the blob and are pointed at directly.
Wavetable base keeps alBnkfNew's semantics (blob value + tbl ROM offset).
The Rare seq table's 8-byte records are likewise rebuilt natively
(portSeqTablePromote).

Then three latent bugs in the stock SGI/Rare audio code, each found by
hardware watchpoint under the deterministic clock:
- reverb.c ring pointers: ALDelay.input/output are u32; negating them as
  array indices relied on 32-bit pointer wrap - natively the pointer
  lands 2^33 past the ring and aSaveBufferImpl's truncation turned that
  into writes over the delay-line sections. Indices now signed.
- music.c set banks via alSeqpSetBank((ALSeqPlayer*)cspPlayer) - Rare's
  own '// typo' comment; the evtq offsets coincided on 32-bit but
  natively the event post scribbled the bank pointer over the player
  node's next link. Implemented the missing alCSPSetBank (posts to the
  CSP's real evtq) and switched the three calls.
- the synth's __allocParam pool sizes nodes as sizeof(ALParam); every
  variant coincided at 28 bytes on 32-bit, but ALStartParamAlt's
  trailing wave pointer outgrows it natively and overhung into the
  adjacent heap objects (the sequence players). Nodes now sized by an
  ALParamNode union of all variants.

Also: music heap doubled on 64-bit (native AL structs) and the memp
permanent pool's 296KB default doubled to hold it (its OOM path is
'while(1);'). 32-bit GOLDENS PASS - byte-identical PCM.

+++ src/libultrare/audio/reverb.c
@@ -92,8 +92,8 @@ Acmd *alFxPull(void *filter, s16 *outp, s32 outCount, s32 sampleOffset,
 
     for (i = 0; i < r->section_count; i++) {
 	d  = &r->delay[i];  /* get the ALDelay structure */
-	in_ptr  = &r->input[-d->input];
-	out_ptr = &r->input[-d->output];
+	in_ptr  = &r->input[-(s32)d->input];
+	out_ptr = &r->input[-(s32)d->output];
 	
 	if (in_ptr == prev_out_ptr) {
 	    SWAP(buff1, buff2);
@@ -265,7 +265,7 @@ Acmd *_loadOutputBuffer(ALFx *r, ALDelay *d, s32 buff, s32 incount, Acmd *p)
          * negative of that as an index into the delay buffer. loadBuffer that uses this
          * value then bumps it up if it is below the  delay buffer.
          */ 
-        out_ptr = &r->input[-(d->output - d->rsdelta)];
+        out_ptr = &r->input[-((s32)d->output - d->rsdelta)];
         ramalign = ((s32)out_ptr & 0x7) >> 1; /* calculate the number of samples needed 
                                                to align the buffer*/
 #ifdef _DEBUG
@@ -295,7 +295,7 @@ Acmd *_loadOutputBuffer(ALFx *r, ALDelay *d, s32 buff, s32 incount, Acmd *p)
         d->rs->first = 0; /* turn off first time flag */
         d->rsdelta += count - incount; /* add the number of samples to d->rsdelta */
     } else {
-        out_ptr = &r->input[-d->output];
+        out_ptr = &r->input[-(s32)d->output];
         ptr = _loadBuffer(r, out_ptr, buff, incount, ptr);
     }
